finshied all schema, models, resolvers incl seacrh & perk. search untested

+++ black_star_mind/graphql/resolvers/perk.js
@@ -15,10 +15,10 @@ const { dateToString } = require('../../helpers/date');
 const { pocketVariables } = require('../../helpers/pocketVars');
 
 module.exports = {
-  perks: async (req) => {
-    // if (!req.isAuth) {
-    //   throw new Error('Unauthenticated!');
-    // }
+  perks: async (args,req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
     try {
       const perks = await Perk.find();
       return perks.map(perk => {
@@ -29,14 +29,14 @@ module.exports = {
     }
   },
   getPerkId: async (args, req) => {
-    // if (!req.isAuth) {
-    //   throw new Error('Unauthenticated!');
-    // }
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
     try {
       const perk = await Perk.findById(args.perkId);
         return {
           ...perk._doc,
-          _id: perk._id,
+          _id: perk.id,
           name: perk.name,
           description: perk.description,
           type: perk.type,
@@ -44,21 +44,22 @@ module.exports = {
           data: perk.data,
           users: perk.users,
           groups: perk.groups,
-          content: perk.content
+          content: perk.content,
+          tags: perk.tags
         };
     } catch (err) {
       throw err;
     }
   },
   getPerkName: async (args, req) => {
-    // if (!req.isAuth) {
-    //   throw new Error('Unauthenticated!');
-    // }
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
     try {
-      const action = await Action.findOne({name: args.name});
+      const perk = await Perk.findOne({name: args.name});
         return {
           ...perk._doc,
-          _id: perk._id,
+          _id: perk.id,
           name: perk.name,
           description: perk.description,
           type: perk.type,
@@ -66,82 +67,138 @@ module.exports = {
           data: perk.data,
           users: perk.users,
           groups: perk.groups,
-          content: perk.content
+          content: perk.content,
+          tags: perk.tags
         };
     } catch (err) {
       throw err;
     }
   },
   getPerkType: async (args, req) => {
-    // if (!req.isAuth) {
-    //   throw new Error('Unauthenticated!');
-    // }
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
     try {
-      const action = await Perk.findOne({type: args.type});
-        return {
-          ...perk._doc,
-          _id: perk._id,
-          name: perk.name,
-          description: perk.description,
-          type: perk.type,
-          subtype: perk.subtype,
-          data: perk.data,
-          users: perk.users,
-          groups: perk.groups,
-          content: perk.content
-        };
+      const perks = await Perk.find({type: args.type});
+      return perks.map(perk => {
+        return transformPerk(perk);
+      });
+        // {
+        // ...perk._doc,
+        // _id: perk._id,
+        // name: perk.name,
+        // description: perk.description,
+        // type: perk.type,
+        // subtype: perk.subtype,
+        // data: perk.data,
+        // users: perk.users,
+        // groups: perk.groups,
+        // content: perk.content,
+        // tags: perk.tags
+        // };
     } catch (err) {
       throw err;
     }
   },
-  // getPerkUser: async (args, req) => {
-    // if (!req.isAuth) {
-    //   throw new Error('Unauthenticated!');
-    // }
-    // try {
-    //   const perks = await Perk.find();
-    //   console.log(perks);
-    // } catch (err) {
-    //   throw(err)
-    // };
-    // try {
-    //   const perk = await Perk.find({ 'users': { $elemMatch: {username: args.userRefInput.username } } });
-      // const perk = await Perk.find({'users.username': {$lte: args.userRefInput.username}});
-      // const perk = await Perk.find({'users.username': args.userRefInput.username});
-  //     console.log(group);
-  //       return {
-  //          ...perk._doc,
-              // _id: perk._id,
-              // name: perk.name,
-              // description: perk.description,
-              // type: perk.type,
-              // subtype: perk.subtype,
-              // data: perk.data,
-              // users: perk.users,
-              // groups: perk.groups,
-              // content: perk.content
-  //       };
-  //   } catch (err) {
-  //     throw err;
-  //   }
-  // },
-  updatePerk: async (args, req) => {
-    // if (!req.isAuth) {
-    //   throw new Error('Unauthenticated!');
-    // }
-    console.log(JSON.stringify(args));
+  getPerkUser: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
     try {
-      const content = await Perk.findOneAndUpdate({_id:args.perkId},{
+      const perks = await Perk.find({'users.username': args.userRefInput.username});
+      return perks.map(perk => {
+        return transformPerk(perk);
+      });
+      // {
+      // ...perk._doc,
+      // _id: perk._id,
+      // name: perk.name,
+      // description: perk.description,
+      // type: perk.type,
+      // subtype: perk.subtype,
+      // data: perk.data,
+      // users: perk.users,
+      // groups: perk.groups,
+      // content: perk.content,
+      // tags: perk.tags
+      // };
+    } catch (err) {
+      throw err;
+    }
+  },
+  getPerkGroup: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
+    try {
+      const perks = await Perk.find({groups: args.groupRefInput});
+      return perks.map(perk => {
+        return transformPerk(perk);
+      });
+      // {
+      // ...perk._doc,
+      // _id: perk._id,
+      // name: perk.name,
+      // description: perk.description,
+      // type: perk.type,
+      // subtype: perk.subtype,
+      // data: perk.data,
+      // users: perk.users,
+      // groups: perk.groups,
+      // content: perk.content,
+      // tags: perk.tags
+      // };
+    } catch (err) {
+      throw err;
+    }
+  },
+  getPerkTag: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
+    try {
+      const perks = await Perk.find({tags: args.tags});
+      return perks.map(perk => {
+        return transformPerk(perk);
+      });
+      // {
+      // ...perk._doc,
+      // _id: perk._id,
+      // name: perk.name,
+      // description: perk.description,
+      // type: perk.type,
+      // subtype: perk.subtype,
+      // data: perk.data,
+      // users: perk.users,
+      // groups: perk.groups,
+      // content: perk.content,
+      // tags: perk.tags
+      // };
+    } catch (err) {
+      throw err;
+    }
+  },
+  updatePerk: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
+    try {
+      const perk = await Perk.findOneAndUpdate({_id:args.perkId},{
         // {
-        type: args.actionInput.type,
-        target: args.actionInput.target,
-        creator: args.actionInput.creator,
-        description: args.actionInput.description,
+        name: args.perkInput.name,
+        description: args.perkInput.description,
+        type: args.perkInput.type,
+        subtype: args.perkInput.subtype,
+        data: args.perkInput.data,
+        users: args.perkInput.users,
+        groups: args.perkInput.groups,
+        content: args.perkInput.content,
+        tags: args.perkInput.tags
       // }
       },{new: true});
         return {
           ...perk._doc,
-          _id: perk._id,
+          _id: perk.id,
           name: perk.name,
           description: perk.description,
           type: perk.type,
@@ -149,22 +206,22 @@ module.exports = {
           data: perk.data,
           users: perk.users,
           groups: perk.groups,
-          content: perk.content
+          content: perk.content,
+          tags: perk.tags
         };
     } catch (err) {
       throw err;
     }
   },
   updatePerkSubtype: async (args, req) => {
-    // if (!req.isAuth) {
-    //   throw new Error('Unauthenticated!');
-    // }
-    console.log(JSON.stringify(args));
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
     try {
-      const perk = await Perk.findOneAndUpdate({_id:args.perkId},{subtype: args.perkSubtypeInput},{new: true});
+      const perk = await Perk.findOneAndUpdate({_id:args.perkId},{subtype:args.perkSubtypeInput},{new: true});
         return {
           ...perk._doc,
-          _id: perk._id,
+          _id: perk.id,
           name: perk.name,
           description: perk.description,
           type: perk.type,
@@ -172,22 +229,22 @@ module.exports = {
           data: perk.data,
           users: perk.users,
           groups: perk.groups,
-          content: perk.content
+          content: perk.content,
+          tags: perk.tags
         };
     } catch (err) {
       throw err;
     }
   },
   updatePerkUser: async (args, req) => {
-    // if (!req.isAuth) {
-    //   throw new Error('Unauthenticated!');
-    // }
-    console.log(JSON.stringify(args));
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
     try {
-      const content = await Perk.findOneAndUpdate({_id:args.perkId},{$push: {users:args.userRefInput}},{new: true});
+      const perk= await Perk.findOneAndUpdate({_id:args.perkId},{$push: {users:args.userRefInput}},{new: true});
         return {
           ...perk._doc,
-          _id: perk._id,
+          _id: perk.id,
           name: perk.name,
           description: perk.description,
           type: perk.type,
@@ -195,22 +252,22 @@ module.exports = {
           data: perk.data,
           users: perk.users,
           groups: perk.groups,
-          content: perk.content
+          content: perk.content,
+          tags: perk.tags
         };
     } catch (err) {
       throw err;
     }
   },
   updatePerkData: async (args, req) => {
-    // if (!req.isAuth) {
-    //   throw new Error('Unauthenticated!');
-    // }
-    console.log(JSON.stringify(args));
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
     try {
-      const content = await Action.findOneAndUpdate({_id:args.perkId},{$push: {data:args.perkDataInput}},{new: true});
+      const perk = await Perk.findOneAndUpdate({_id:args.perkId},{$push: {data:args.perkDataInput}},{new: true});
         return {
           ...perk._doc,
-          _id: perk._id,
+          _id: perk.id,
           name: perk.name,
           description: perk.description,
           type: perk.type,
@@ -218,22 +275,22 @@ module.exports = {
           data: perk.data,
           users: perk.users,
           groups: perk.groups,
-          content: perk.content
+          content: perk.content,
+          tags: perk.tags
         };
     } catch (err) {
       throw err;
     }
   },
   updatePerkContent: async (args, req) => {
-    // if (!req.isAuth) {
-    //   throw new Error('Unauthenticated!');
-    // }
-    console.log(JSON.stringify(args));
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
     try {
-      const group = await Perk.findOneAndUpdate({_id:args.perkId},{$push: {content:args.contentRefInput}},{new: true});
+      const perk = await Perk.findOneAndUpdate({_id:args.perkId},{$push: {content:args.contentRefInput}},{new: true});
         return {
           ...perk._doc,
-          _id: perk._id,
+          _id: perk.id,
           name: perk.name,
           description: perk.description,
           type: perk.type,
@@ -241,22 +298,45 @@ module.exports = {
           data: perk.data,
           users: perk.users,
           groups: perk.groups,
-          content: perk.content
+          content: perk.content,
+          tags: perk.tags
+        };
+    } catch (err) {
+      throw err;
+    }
+  },
+  updatePerkGroup: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
+    try {
+      const perk = await Perk.findOneAndUpdate({_id:args.perkId},{$push: {groups:args.groupRefInput}},{new: true});
+        return {
+          ...perk._doc,
+          _id: perk.id,
+          name: perk.name,
+          description: perk.description,
+          type: perk.type,
+          subtype: perk.subtype,
+          data: perk.data,
+          users: perk.users,
+          groups: perk.groups,
+          content: perk.content,
+          tags: perk.tags
         };
     } catch (err) {
       throw err;
     }
   },
   updatePerkTag: async (args, req) => {
-    // if (!req.isAuth) {
-    //   throw new Error('Unauthenticated!');
-    // }
-    console.log(JSON.stringify(args));
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
     try {
-      const content = await Perk.findOneAndUpdate({_id:args.perkId},{$push: {tags:args.tags}},{new: true});
+      const perk = await Perk.findOneAndUpdate({_id:args.perkId},{$push: {tags:args.tags}},{new: true});
         return {
           ...perk._doc,
-          _id: perk._id,
+          _id: perk.id,
           name: perk.name,
           description: perk.description,
           type: perk.type,
@@ -272,55 +352,64 @@ module.exports = {
     }
   },
   deletePerk: async (args, req) => {
-    // if (!req.isAuth) {
-    //   throw new Error('Unauthenticated!');
-    // }
-    // console.log(JSON.stringify(args));
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
     try {
       const perk = await Perk.findByIdAndRemove(args.perkId);
         return {
           ...perk._doc,
-          _id: perk.id
+          _id: perk.id,
+          name: perk.name,
+          description: perk.description,
+          type: perk.type,
+          subtype: perk.subtype,
+          data: perk.data,
+          users: perk.users,
+          groups: perk.groups,
+          content: perk.content,
+          tags: perk.tags
         };
     } catch (err) {
       throw err;
     }
   },
-  createPerk: async (args) => {
-    // if (!req.isAuth) {
-    //   throw new Error('Unauthenticated!');
-    // }
-    console.log(JSON.stringify(args));
+  createPerk: async (args,req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!');
+    }
     try {
-      // const existingContent = await Action.findOne({ title: args.actionInput.title });
-      // if (existingContent) {
-      //   throw new Error('Title exists already.');
-      // }
+      const existingPerk = await Perk.findOne({ name: args.perkInput.name });
+      if (existingPerk) {
+        throw new Error('perk name exists already.');
+      }
       const perk = new Perk({
         name: args.perkInput.name,
         description: args.perkInput.description,
-        type: args.perkInput.type
+        type: args.perkInput.type,
+        subtype: args.perkInput.subtype,
+        data: args.perkInput.data,
+        users: args.perkInput.users,
+        groups: args.perkInput.groups,
+        content: args.perkInput.content,
+        tags: args.perkInput.tags
       });
 
       const result = await perk.save();
 
       return {
-        ...perk._doc,
-        _id: perk._id,
-        name: perk.name,
-        description: perk.description,
-        type: perk.type,
-        subtype: perk.subtype,
-        data: perk.data,
-        users: perk.users,
-        groups: perk.groups,
-        content: perk.content,
-        tags: perk.tags
+        ...result._doc,
+        _id: result.id,
+        name: result.name,
+        description: result.description,
+        type: result.type,
+        subtype: result.subtype,
+        data: result.data,
+        users: result.users,
+        groups: result.groups,
+        content: result.content,
+        tags: result.tags
       };
-      // pocketVariables.key01 = "contentId";
-      // pocketVariables.value01 = result.id;
-      // return pocketVariables;
-      // console.log("pocket vars:  " + pocketVariables);
     } catch (err) {
       throw err;
     }
